JWT: Authenticaltion and Authorization for APIs

+++ controllers/api/v1/post_api.js
@@ -25,17 +25,19 @@ module.exports.destroy = async function(req,res){
     try {
          //Find the post having the id passed in URL
         let post =  await Post.findById(req.params.id);
+         if(post.user == req.user.id){
             post.remove();
             //Delete all those comments which have post: req.params.id
             await Comment.deleteMany({post: req.params.id });
-
             return res.json(200, {
                 message: "Post and associated comments deleted"
-            })
-        
-        // else{
-        //     return res.redirect('back');
-        // }
+            });
+         }
+        else{
+            return res.json(401,{
+                message: "You can not delete this post!"
+            });
+        }
     } catch (error) {
         return res.json(500, {
             message: "Internal Server Error"
